New terrain generation design
